Draw out new DistanceUnit value object class
service, for working with distance units, which were merely stored in Distance objects as
a symbol (e.g ':km'). That was clunky, as the tests began to show me. So this refactor
pulls out DistanceUnit as a class in its own right, which encapsulates the Unit of
Mesasure (UOM) symbol, description, and conversion factor (to KM). I may also extract the
remaining class level methods into a kind of units repository, which will also make it
easier for users to load up custom UOMs.