Begin marrying the new Distance and Pace classes with the rest of the domain
Summary:
I created the Distance and Pace classes with the units baked in, but I still needed to actually use it in the rest of the domain. Previously I just used a number for a distance; units had to be passed in along with it. Also, paces were only expressed as a RunbyTime; units were passed in separately. That was clunky. The only thing I don't like now is all of the new'ing up and 'is_?' Class queries. I'll need to clean this up.