Generate version number in Rake task
Adds a rake task called gen_version_number, which runs together with the default rake tasks. This task creates the version number by finding the latest annotated "v" tag and appending the number of commits to that as the "teeny" portion of the version number.

Once the version number is generated, it is added to a generated file called `version.g.rb`, which is not under version control.

Why all of this? I don't want to manually increment the version number every time I commit. This method removes a runtime dependency on Git for determining the version number.